Replaced ZzdReader with getInputFile (same for writer/output)
I really struggled with producing the ZzdReader and ZzdWriter structs,
and I suppose I'm glad I did it for the practice. But in hindsight,
clearly stdin and stdout are 'files' that can be returned just like any
other, eliminating the need for returning readers and writers and
allocating memory.

Not sure how I forgot that simple fact, but learning from mistakes is
what this project is all about.